rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG backend"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "bcrypt >=4.3.0,<5.0.0",
    "fastapi >=0.115.12,<0.116.0",
    "gunicorn >=23.0.0,<24.0.0",
    "passlib >=1.7.4,<2.0.0",
    "requests >=2.32.3,<3.0.0",
    "tortoise-orm >=0.25.0,<0.26.0",
    "uvicorn >=0.34.2,<0.35.0",
    "aerich >=0.9.0,<0.10.0",
    "pillow >=11.2.1,<12.0.0",
    "schedule >=1.2.2,<2.0.0",
    "haystack-ai ==2.10.0",
    "python-docx >=1.1.2,<2.0.0",
    "markdown-it-py >=3.0.0,<4.0.0",
    "mdit-plain >=1.0.1,<2.0.0",
    "pypdf >=5.5.0,<6.0.0",
    "colorama >=0.4.6,<0.5.0",
    "pyyaml >=6.0.2,<7.0.0",
    "pyjwt >=2.10.1,<3.0.0",
    "librosa >=0.11.0,<0.12.0",
    "jinja2 >=3.1.6,<4.0.0",
    "openpyxl >=3.1.5,<4.0.0",
    "tabulate >=0.9.0,<0.10.0",
    "pynvml >=12.0.0,<13.0.0",
    "python-multipart >=0.0.20,<0.0.21",
    "markdownify >=1.1.0,<2.0.0",
    "openai >=1.82.0,<2.0.0",
    "python-jose >=3.4.0,<4.0.0",
    "ply >=3.11,<4.0",
    "aiohttp >=3.11.18,<4.0.0",
    "matplotlib >=3.10.3,<4.0.0",
    "milvus-haystack >=0.0.16,<0.0.17",
    "aiostream >=0.6.4,<0.7.0",
    "pymilvus >=2.5.9,<3.0.0",
    "xlrd >=2.0.1,<3.0.0",
    "xlwings >=0.33.15,<0.34.0",
    "ipython >=9.2.0,<10.0.0",
    "mcp >=1.9.1,<2.0.0",
]

[tool.setuptools]
packages = ["core", "server"]
package-dir = { "" = "src" }

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"